+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 ## [Unreleased]
 
 - Work in progress...
+
+## [v0.0.8] - 2016-07-28
+
 - added *HTML5 Asteroids** from <https://github.com/dmcinnes/HTML5-Asteroids>
 - changed `candyjam` to real name `dental-defender`
 
@@ -48,7 +51,8 @@
 
 - Repo created: <https://github.com/attogram/games>
 
-[Unreleased]: https://github.com/attogram/games/compare/v0.0.7...HEAD
+[Unreleased]: https://github.com/attogram/games/compare/v0.0.8...HEAD
+[v0.0.8]: https://github.com/attogram/games/compare/v0.0.7...v0.0.8
 [v0.0.7]: https://github.com/attogram/games/compare/v0.0.6...v0.0.7
 [v0.0.6]: https://github.com/attogram/games/compare/v0.0.5...v0.0.6
 [v0.0.5]: https://github.com/attogram/games/compare/v0.0.4...v0.0.5